Delete SpotlessTaskModern, and move the important parts over to SpotlessTaskModern.
